move OSSL setup functions to the new osslsetup package
