+++ openssl.go
@@ -7,37 +7,27 @@ import (
 	"errors"
 	"math/bits"
 	"strconv"
-	"sync"
 	"unsafe"
 
 	"github.com/golang-fips/openssl/v2/internal/ossl"
+	"github.com/golang-fips/openssl/v2/osslsetup"
 )
 
 var (
 	// vMajor and vMinor hold the major/minor OpenSSL version.
 	// It is only populated if Init has been called.
-	vMajor, vMinor, vPatch uint
+	vMajor, vMinor, vPatch int
 )
-
-var (
-	initOnce sync.Once
-	initErr  error
-)
-
-var isBigEndian bool
 
 // CheckVersion checks if the OpenSSL version can be loaded
 // and if the FIPS mode is enabled.
 // This function can be called before Init.
 // All OpenSSL functions used in here should be tagged with "init_1" or "init_3" in shims.h.
 func CheckVersion(version string) (exists, fips bool) {
-	close, err := initForCheckVersion(version)
-	if err != nil {
-		return false, false
-	}
-	defer close()
-	return true, FIPS()
+	return osslsetup.CheckVersion(version)
 }
+
+var isBigEndian bool
 
 // Init loads and initializes OpenSSL from the shared library at path.
 // It must be called before any other OpenSSL call, except CheckVersion.
@@ -49,24 +39,25 @@ func CheckVersion(version string) (exists, fips bool) {
 // For example, `file=libcrypto.so.1.1.1k-fips` makes Init look for the shared
 // library libcrypto.so.1.1.1k-fips.
 func Init(file string) error {
-	initOnce.Do(func() {
-		buf := [2]byte{}
-		*(*uint16)(unsafe.Pointer(&buf[0])) = uint16(0xABCD)
+	buf := [2]byte{}
+	*(*uint16)(unsafe.Pointer(&buf[0])) = uint16(0xABCD)
 
-		switch buf {
-		case [2]byte{0xCD, 0xAB}:
-			isBigEndian = false
-		case [2]byte{0xAB, 0xCD}:
-			isBigEndian = true
-		default:
-			panic("Could not determine native endianness.")
-		}
-		initErr = opensslInit(file)
-	})
-	return initErr
+	switch buf {
+	case [2]byte{0xCD, 0xAB}:
+		isBigEndian = false
+	case [2]byte{0xAB, 0xCD}:
+		isBigEndian = true
+	default:
+		panic("Could not determine native endianness.")
+	}
+	if err := osslsetup.Init(file); err != nil {
+		return err
+	}
+	vMajor, vMinor, vPatch = osslsetup.Version()
+	return nil
 }
 
-func utoa(n uint) string {
+func utoa(n int) string {
 	return strconv.FormatUint(uint64(n), 10)
 }
 
@@ -75,7 +66,7 @@ func errUnsupportedVersion() error {
 }
 
 // checkMajorVersion panics if the current major version is not expected.
-func checkMajorVersion(expected uint) {
+func checkMajorVersion(expected int) {
 	if vMajor != expected {
 		panic("openssl: incorrect major version (" + strconv.Itoa(int(vMajor)) + "), expected " + strconv.Itoa(int(expected)))
 	}
@@ -86,137 +77,37 @@ type fail string
 func (e fail) Error() string { return "openssl: " + string(e) + " failed" }
 
 // VersionText returns the version text of the OpenSSL currently loaded.
+//
+//go:fix inline
 func VersionText() string {
-	return goString(ossl.OpenSSL_version(0))
+	return osslsetup.VersionText()
 }
 
 // FIPS returns true if OpenSSL is running in FIPS mode and there is
 // a provider available that supports FIPS. It returns false otherwise.
 // All OpenSSL functions used in here should be tagged with "init_1" or "init_3" in shims.h.
+//
+//go:fix inline
 func FIPS() bool {
-	switch vMajor {
-	case 1:
-		return ossl.FIPS_mode() == 1
-	case 3:
-		// Check if the default properties contain `fips=1`.
-		if ossl.EVP_default_properties_is_fips_enabled(nil) != 1 {
-			// Note that it is still possible that the provider used by default is FIPS-compliant,
-			// but that wouldn't be a system or user requirement.
-			return false
-		}
-		// Check if the SHA-256 algorithm is available. If it is, then we can be sure that there is a provider available that matches
-		// the `fips=1` query. Most notably, this works for the common case of using the built-in FIPS provider.
-		//
-		// Note that this approach has a small chance of false negative if the FIPS provider doesn't provide the SHA-256 algorithm,
-		// but that is highly unlikely because SHA-256 is one of the most common algorithms and fundamental to many cryptographic operations.
-		// It also has a small chance of false positive if the FIPS provider implements the SHA-256 algorithm but not the other algorithms
-		// used by the caller application, but that is also unlikely because the FIPS provider should provide all common algorithms.
-		return proveSHA256("")
-	default:
-		panic(errUnsupportedVersion())
-	}
+	return osslsetup.FIPS()
 }
 
 // FIPSCapable returns true if the provider used by default matches the `fips=yes` query.
-// It is useful for checking whether OpenSSL is capable of running in FIPS mode regardless
-// of whether FIPS mode is explicitly enabled. For example, Azure Linux 3 doesn't set the
-// `fips=yes` query in the default properties, but sets the default provider to be SCOSSL,
-// which is FIPS-capable.
+// See [osslsetup.FIPSCapable] for details.
 //
-// Considerations:
-//   - Multiple calls to FIPSCapable can return different values if [SetFIPS] is called in between.
-//   - Can return true even if [FIPS] returns false, because [FIPS] also checks whether
-//     the default properties contain `fips=yes`.
-//   - When using OpenSSL 3, will always return true if [FIPS] returns true.
-//   - When using OpenSSL 1, Will always return the same value as [FIPS].
-//   - OpenSSL 3 doesn't provide a way to know if a provider is FIPS-capable. This function uses
-//     some heuristics that should be treated as an implementation detail that may change in the future.
+//go:fix inline
 func FIPSCapable() bool {
-	if FIPS() {
-		return true
-	}
-	if vMajor == 3 {
-		// Load the provider with and without the `fips=yes` query.
-		// If the providers are the same, then the default provider is FIPS-capable.
-		provFIPS := sha256Provider(_ProviderNameFips)
-		if provFIPS == nil {
-			return false
-		}
-		provDefault := sha256Provider("")
-		return provFIPS == provDefault
-	}
-	return false
-}
-
-// isProviderAvailable checks if the provider with the given name is available.
-// This function is used in export_test.go, but must be defined here as test files can't access C functions.
-func isProviderAvailable(name string) bool {
-	if vMajor == 1 {
-		return false
-	}
-	return ossl.OSSL_PROVIDER_available(nil, unsafe.StringData(name+"\x00")) == 1
+	return osslsetup.FIPSCapable()
 }
 
 // SetFIPS enables or disables FIPS mode.
 //
 // For OpenSSL 3, if there is no provider available that supports FIPS mode,
 // SetFIPS will try to load a built-in provider that supports FIPS mode.
+//
+//go:fix inline
 func SetFIPS(enable bool) error {
-	if FIPS() == enable {
-		// Already in the desired state.
-		return nil
-	}
-	var mode int32
-	if enable {
-		mode = int32(1)
-	} else {
-		mode = int32(0)
-	}
-	switch vMajor {
-	case 1:
-		if _, err := ossl.FIPS_mode_set(mode); err != nil {
-			return err
-		}
-		return nil
-	case 3:
-		var shaProps, provName cString
-		if enable {
-			shaProps = _PropFIPSYes
-			provName = _ProviderNameFips
-		} else {
-			shaProps = _PropFIPSNo
-			provName = _ProviderNameDefault
-		}
-		if !proveSHA256(shaProps) {
-			// There is no provider available that supports the desired FIPS mode.
-			// Try to load the built-in provider associated with the given mode.
-			if p, _ := ossl.OSSL_PROVIDER_try_load(nil, provName.ptr(), 1); p == nil {
-				// The built-in provider was not loaded successfully, we can't enable FIPS mode.
-				return errors.New("openssl: FIPS mode not supported by any provider")
-			}
-		}
-		_, err := ossl.EVP_default_properties_enable_fips(nil, mode)
-		return err
-	default:
-		panic(errUnsupportedVersion())
-	}
-}
-
-// sha256Provider returns the provider for the SHA-256 algorithm
-// using the given properties.
-func sha256Provider(props cString) ossl.OSSL_PROVIDER_PTR {
-	md, _ := ossl.EVP_MD_fetch(nil, _DigestNameSHA2_256.ptr(), props.ptr())
-	if md == nil {
-		return nil
-	}
-	defer ossl.EVP_MD_free(md)
-	return ossl.EVP_MD_get0_provider(md)
-}
-
-// proveSHA256 checks if the SHA-256 algorithm is available
-// using the given properties.
-func proveSHA256(props cString) bool {
-	return sha256Provider(props) != nil
+	return osslsetup.SetFIPS(enable)
 }
 
 var zero byte
@@ -344,7 +235,7 @@ func bnToBinPad(bn ossl.BIGNUM_PTR, to []byte) error {
 // versionAtOrAbove returns true when
 // (vMajor, vMinor, vPatch) >= (major, minor, patch),
 // compared lexicographically.
-func versionAtOrAbove(major, minor, patch uint) bool {
+func versionAtOrAbove(major, minor, patch int) bool {
 	return vMajor > major || (vMajor == major && vMinor > minor) || (vMajor == major && vMinor == minor && vPatch >= patch)
 }
