+++ osslsetup/init.go
@@ -1,6 +1,6 @@
 //go:build !cmd_go_bootstrap && (cgo || goexperiment.ms_nocgo_opensslcrypto)
 
-package openssl
+package osslsetup
 
 import (
 	"errors"
